+++ .gitignore
@@ -4,6 +4,6 @@ yarn-error.log
 .env
 *.iml
 .idea
-types/ethers-contracts
-dist
-build
+/types/ethers-contracts
+/dist
+/build
